set up chopped prefab....
so that only for the chopped mini game, the bottom bar of the play area is set to isTrigger = true.  This is what is need so the balls pass through the bottom barrier.  When switching the play area to the prefab, it was set to istrigger=false which was causing the balls to not pass through